G9-api/PurchasesService/verCarritoCompras/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
psycopg2==2.9.6
psycopg2_binary==2.9.6
PyJWT==2.6.0
PyJWT==2.6.0
python-dotenv==1.0.0
waitress==2.1.2
